Added Thumbnail icons and pictures from Blizzard api to character data model. Update Character list view and controller. Updated Character View Added NewsFeed using WoWHead Api Many layout updated in all views. Added Items tooltips using WoWHead Api. A lot of styles were updated in style.css

+++ WoWJunkyard/WoWJunkyard/Data/Models/AzeriteItem.cs
@@ -6,7 +6,7 @@ namespace WoWJunkyard.Data.Models
     public class AzeriteItem
     {
         [Key]
-        public int AzeriteItemId { get; set; }
+        public long AzeriteItemId { get; set; }
 
         public long Id { get; set; }
